Move SST test file into standard_plugins/tests
